+++ app/Http/Controllers/Api/V1/Admin/PermissionsApiController.php
@@ -5,34 +5,52 @@ namespace App\Http\Controllers\Api\V1\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StorePermissionRequest;
 use App\Http\Requests\UpdatePermissionRequest;
+use App\Http\Resources\Admin\PermissionResource;
 use App\Permission;
+use Gate;
+use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\Response;
 
 class PermissionsApiController extends Controller
 {
     public function index()
     {
-        $permissions = Permission::all();
+        abort_if(Gate::denies('permission_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        return $permissions;
+        return new PermissionResource(Permission::all());
     }
 
     public function store(StorePermissionRequest $request)
     {
-        return Permission::create($request->all());
-    }
+        $permission = Permission::create($request->all());
 
-    public function update(UpdatePermissionRequest $request, Permission $permission)
-    {
-        return $permission->update($request->all());
+        return (new PermissionResource($permission))
+            ->response()
+            ->setStatusCode(Response::HTTP_CREATED);
     }
 
     public function show(Permission $permission)
     {
-        return $permission;
+        abort_if(Gate::denies('permission_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        return new PermissionResource($permission);
+    }
+
+    public function update(UpdatePermissionRequest $request, Permission $permission)
+    {
+        $permission->update($request->all());
+
+        return (new PermissionResource($permission))
+            ->response()
+            ->setStatusCode(Response::HTTP_ACCEPTED);
     }
 
     public function destroy(Permission $permission)
     {
-        return $permission->delete();
+        abort_if(Gate::denies('permission_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        $permission->delete();
+
+        return response(null, Response::HTTP_NO_CONTENT);
     }
 }
