+++ app/Http/Controllers/Api/V1/Admin/RolesApiController.php
@@ -5,34 +5,54 @@ namespace App\Http\Controllers\Api\V1\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreRoleRequest;
 use App\Http\Requests\UpdateRoleRequest;
+use App\Http\Resources\Admin\RoleResource;
 use App\Role;
+use Gate;
+use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\Response;
 
 class RolesApiController extends Controller
 {
     public function index()
     {
-        $roles = Role::all();
+        abort_if(Gate::denies('role_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        return $roles;
+        return new RoleResource(Role::with(['permissions'])->get());
     }
 
     public function store(StoreRoleRequest $request)
     {
-        return Role::create($request->all());
-    }
+        $role = Role::create($request->all());
+        $role->permissions()->sync($request->input('permissions', []));
 
-    public function update(UpdateRoleRequest $request, Role $role)
-    {
-        return $role->update($request->all());
+        return (new RoleResource($role))
+            ->response()
+            ->setStatusCode(Response::HTTP_CREATED);
     }
 
     public function show(Role $role)
     {
-        return $role;
+        abort_if(Gate::denies('role_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        return new RoleResource($role->load(['permissions']));
+    }
+
+    public function update(UpdateRoleRequest $request, Role $role)
+    {
+        $role->update($request->all());
+        $role->permissions()->sync($request->input('permissions', []));
+
+        return (new RoleResource($role))
+            ->response()
+            ->setStatusCode(Response::HTTP_ACCEPTED);
     }
 
     public function destroy(Role $role)
     {
-        return $role->delete();
+        abort_if(Gate::denies('role_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        $role->delete();
+
+        return response(null, Response::HTTP_NO_CONTENT);
     }
 }
